pyproject.toml
[project]
name = "otc"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "shyu61", email = "ohs6261@gmail.com" }
]
dependencies = [
    "pandas>=2.0.3",
    "polars>=0.19.19",
    "numpy>=1.24.4",
    "seaborn>=0.13.0",
    "matplotlib>=3.7.4",
    "scikit-learn>=1.3.2",
    "ipykernel>=6.27.1",
    "pyarrow>=14.0.1",
    "tqdm>=4.66.1",
    "kaggle>=1.5.16",
    "lightgbm>=4.1.0",
    "xgboost>=2.0.2",
    "catboost>=1.2.2",
    "hydra-core>=1.3.2",
    "joblib>=1.3.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
